fix(build): extract computeWaves to shared lib so epics route can import it
